+++ pysys-examples/internal/testcases/PySys_internal_083/run.py
@@ -14,9 +14,9 @@ class PySysTest(BaseTest):
 		self.log.info('Python local/default/preferred encoding is %s; will test with non-local encoding %s', locale.getpreferredencoding(), TEST_ENCODING)
 		if locale.getpreferredencoding() in ['ANSI_X3.4-1968', 'ascii']: self.skipTest('cannot run in ASCII locale')
 
-		with io.open(self.output+'/test-nonlocal.txt', 'w', encoding=TEST_ENCODING) as f:
-			f.write(os.linesep.join([TEST_STR, TEST_STR, 'otherstring']))
 		self.__myDefaultEncoding = None
+
+		self.write_text('test-nonlocal.txt', os.linesep.join([TEST_STR, TEST_STR, 'otherstring']), encoding=TEST_ENCODING)
 
 	def validate(self):
 		if TEST_ENCODING == 'utf-8': # can't even read the file without an exception on utf8 systems